- Avoid using fmt.Sprintf
- Avoid allocating intermediate slices
- Write directly to the buffer during iterating instead of strings.Join

before
```
goos: linux
goarch: amd64
pkg: github.com/Masterminds/squirrel
cpu: 11th Gen Intel(R) Core(TM) i3-1115G4 @ 3.00GHz
BenchmarkPlaceholdersArray-4            58258798                17.99 ns/op           18 B/op          0 allocs/op
BenchmarkPlaceholdersStrings-4          1000000000               0.4272 ns/op          2 B/op          0 allocs/op
BenchmarkSquirrelSelectSimple-4           335403              3482 ns/op            2472 B/op         49 allocs/op
BenchmarkSquirrelSelectConditional-4      218160              5563 ns/op            3708 B/op         78 allocs/op
BenchmarkSquirrelSelectComplex-4           63566             19078 ns/op           12362 B/op        275 allocs/op
BenchmarkSquirrelSelectSubquery-4          82434             15066 ns/op            9577 B/op        205 allocs/op
BenchmarkSquirrelSelectMoreComplex-4       46256             25689 ns/op           17218 B/op        391 allocs/op
BenchmarkSquirrelInsert-4                 292023              4127 ns/op            3016 B/op         67 allocs/op
BenchmarkSquirrelUpdateSetColumns-4       183298              6639 ns/op            4376 B/op        101 allocs/op
BenchmarkSquirrelUpdateSetMap-4           173980              6824 ns/op            4448 B/op        103 allocs/op
BenchmarkSquirrelDelete-4                 289329              4167 ns/op            2536 B/op         61 allocs/op
PASS
ok      github.com/Masterminds/squirrel 14.236s
```

after
```
goos: linux
goarch: amd64
pkg: github.com/Masterminds/squirrel
cpu: 11th Gen Intel(R) Core(TM) i3-1115G4 @ 3.00GHz
BenchmarkPlaceholdersArray-4            93750074                11.69 ns/op           18 B/op          0 allocs/op
BenchmarkPlaceholdersStrings-4          1000000000               0.3634 ns/op          2 B/op          0 allocs/op
BenchmarkSquirrelSelectSimple-4           348493              3390 ns/op            2472 B/op         49 allocs/op
BenchmarkSquirrelSelectConditional-4      222134              5405 ns/op            3708 B/op         78 allocs/op
BenchmarkSquirrelSelectComplex-4           67896             17862 ns/op           12120 B/op        260 allocs/op
BenchmarkSquirrelSelectSubquery-4          87997             13688 ns/op            9384 B/op        193 allocs/op
BenchmarkSquirrelSelectMoreComplex-4       49639             24260 ns/op           16800 B/op        367 allocs/op
BenchmarkSquirrelInsert-4                 307369              3925 ns/op            2824 B/op         61 allocs/op
BenchmarkSquirrelUpdateSetColumns-4       202728              5952 ns/op            4136 B/op         88 allocs/op
BenchmarkSquirrelUpdateSetMap-4           195135              6164 ns/op            4208 B/op         90 allocs/op
BenchmarkSquirrelDelete-4                 299464              3984 ns/op            2536 B/op         61 allocs/op
PASS
ok      github.com/Masterminds/squirrel 14.009s
```