@andersfylling 2021-11-27
I suggest running each multiple times, say 10 times, to figure out error margins / p value.

You can then use benchstat to show the delta for improvement: 
```bash
go test -bench=. -count 10 >old.txt
go test -bench=. -count 10 >new.txt
benchstat old.txt new.txt
```

example output:
```
$ benchstat old.txt new.txt
name        old time/op  new time/op  delta
GobEncode   13.6ms ± 1%  11.8ms ± 1%  -13.31% (p=0.016 n=4+5)
JSONEncode  32.1ms ± 1%  31.8ms ± 1%     ~    (p=0.286 n=4+5)
```

https://pkg.go.dev/golang.org/x/perf/cmd/benchstat#section-readme